Move Metrics to Performance Module
Metrics are strongly related to performance, as they are calculated in
the scope of the CPM
