+++ spec/domain/performance/metrics/zero_page_views_spec.rb
@@ -1,0 +1,17 @@
+module Performance
+  RSpec.describe Metrics::ZeroPageViews do
+    subject { Metrics::ZeroPageViews }
+
+    let!(:content_items) {
+      [
+        create(:content_item, one_month_page_views: 1),
+        create(:content_item, one_month_page_views: 0),
+        create(:content_item, one_month_page_views: 0)
+      ]
+    }
+
+    it "returns the number of items with zero page views in the collection" do
+      expect(subject.new(ContentItem.all).run).to eq(zero_page_views: { value: 2 })
+    end
+  end
+end
